handler: add setup/teardown functions for handlers
Before starting an OM run the setup is called for every handler, and
then for every plugin enabled the handlers have the setup_for_plugin
called, allowing them to prepare to receive data from those plugins.
For database handlers this is important to allow tables to be
created before receiving the data.

After an OM run is done, the teardown function is called, allowing
the handlers to do some final cleanup if needed.